Refactored loggers part into loggers.go.
